requirements.txt
pyspark==3.4.1
pandas
streamlit
matplotlib
plotly
python-dotenv
google-generativeai